requirements-dev.txt
# Testing framework
pytest==8.3.4
pytest-cov==5.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2
pre-commit==3.8.0

# Test utilities
Faker==27.2.0
freezegun==1.5.2

# Coverage reporting
coverage==7.6.12

# HTTP client for API testing
requests==2.32.3

# Mock database for unit tests
testing.mysql==1.5.1